Fix writing to debug log file on some systems
This fixes the usage of variadic arguments in some functions
in helper.cpp (for example __pdprint).
The va_list was forwarded to multiple functions (vw_printw and vfprintf),
which is not actually allowed and only worked by chance on some
implementations. On linux+gcc for example the debug log would get
created, but nothing got written to it. The solution is to call va_start
multiple times (once for each usage), or copy the va_list with va_copy.