Thread synchronization for terminal, and debug communication
**PC Side:**

* Multiple threads can read/modify the terminal `local_mesgqueue` at once
* Multiple threads can read/modify the debug `local_mesgqueue` at once

I was seeing crashes because of this, but it's possible for more subtle
effects such as corrupt data. Fixed by protecting access using mutexes.

**Flashcart Side (Libultra Only):**

The debugger code relies on the USB thread having the highest priority so
other threads don't interrupt it during reads/writes. However, it is possible
for the USB thread to block in `usb_dma_write()` when waiting for the DMA to
complete. During this time, another thread (such as the RDB thread) can call
USB related functions and corrupt the internal state.

I was seeing invalid GDB responses (contents from multiple messages being
sent back as one). Fixed with the same workaround as in `usb_dma_read()`:
use a busy loop to wait for the DMA to complete. This could also be solved
using `osSendMesg()` / `osRecvMesg()`.